Fix opacity

+++ cartoframes/viz/styles/size_continuous_style.py
@@ -20,7 +20,7 @@ def size_continuous_style(
           '[2, 40]' for point geometries and '[1, 10]' for lines.
         color (str, optional): hex, rgb or named color value.
           Defaults is '#FFB927' for point geometries and '#4CC8A3' for lines.
-        opacity (int, optional): Opacity value for point color and line features.
+        opacity (float, optional): Opacity value for point color and line features.
           Default is '0.8'.
         stroke_color (str, optional): Color of the stroke on point features.
           Default is '#222'.
